## Added Rust Support to Flash ORM

At the moment, the Rust implementation is still limited and should be considered an initial release. In future updates, we plan to expand its capabilities by adding more database drivers, supporting additional databases, and improving the overall developer experience.

### Currently Supported Databases

* SQLite
* PostgreSQL
* MySQL

### Current Limitations

* CQL databases are not yet supported.
* SQLx is currently the only supported Rust database driver.

SQLx provides a solid foundation for the initial release, and support for additional Rust database drivers will be added in future updates.

+++ example/java/java/CountUsersRow.java
@@ -4,7 +4,7 @@ package java;
 
 
 public record CountUsersRow(
-    int total_users,
-    int admin_count,
-    int regular_count
+    long total_users,
+    long admin_count,
+    long regular_count
 ) {}
